backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
supabase>=1.0.0
google-genai>=0.13.0
requests>=2.31.0
pydantic>=1.10.12
pymongo
bcrypt
PyJWT
